requirements.txt
scikit-learn
mysql-connector-python
joblib
